safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEMorphologyElement-svgdom-operator-prop-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: SVGFEMorphologyElement
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Erode radius = 4
Tests dynamic updates of the 'operator' property of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
